AuthGroup:
-Adding of UpdateGroup and removeGroup
-Change of getAllGroups to see the associated users of the group
-Adition of UserDTO and GroupWithUsersDTO

+++ src/main/java/com/RulesTV/RulesTV/services/AuthTokenService.java
@@ -1,5 +1,4 @@
 package com.RulesTV.RulesTV.services;
-
 import com.RulesTV.RulesTV.entity.AuthToken;
 import com.RulesTV.RulesTV.repositories.AuthTokenRepository;
 import com.RulesTV.RulesTV.repositories.UserRepository;
@@ -19,11 +18,14 @@ public class AuthTokenService {
         this.userRepository = userRepository;
     }
 
+
     public List<AuthToken> getAllTokens() {
+        AuthUserPermissionService.checkSuperAdminAccess();
         return tokenRepository.findAll();
     }
 
     public List<AuthToken> getTokensByUserId(Integer userId){
+        AuthUserPermissionService.checkSuperAdminAccess();
         return tokenRepository.findByUserId(userId);
     }
 
@@ -32,6 +34,7 @@ public class AuthTokenService {
     }
 
     public boolean deleteToken(String token) {
+        AuthUserPermissionService.checkSuperAdminAccess();
         Optional<AuthToken> authTokenOptional = tokenRepository.findByToken(token);
         if (authTokenOptional.isPresent()) {
             tokenRepository.deleteByToken(token);
